Added initial scaffold for Syft client with relevant methods.
